Workaround go-libzfs/libzfs bug
When creating multiple times a dataset prefixed by the same pool name,
but on different pools (iterating to the next use case), all datasets
always refer to the first pool.
This creates issues as the mountpath isn't the correct one.

For now, workaround by prefixing the pool name with the unique test
name.

Co-authored-by: Jean-Baptiste Lallement <jibel@ubuntu.com>