Add ordering to tests
Our tests needs ordering (especially with -update): the output of one is
the input of the next one.
We could have avoided the lock when -update isn't specified, it seems
though that zfs doesn't like heavy parallel workload on the machine and
we get random one-time failure due to zfs module not actionning the
expected state (pool not loaded).
Still use a timeout with tests are filtered, if the other tests didn't
start in $time.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>