Fix grub-probe mock calling the real one
In TestBootlist, we are calling the real grub-probe to match the
underlying device.
In TestGrubMkConfig, we are using the mock grub-probe for the whole stack,
and thus, not matchin the underlying device.
Now, fallback for the only call made by TestBootlist to the real
grub-probe binary. This will be called again when generating the grub
menu in the 3rd step, but will be a no-op.

Thanks to this, we can reuse the same grubmenu reference files.

Co-authored-by: Jean-Baptiste Lallement <jean-baptiste@ubuntu.com>